+++ src/project/infrastructure/postgres/models.py
@@ -1,0 +1,13 @@
+from sqlalchemy.orm import Mapped, mapped_column
+
+from project.infrastructure.postgres.database import Base
+
+
+class Client(Base):
+    __tablename__ = "clients"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[str] = mapped_column(nullable=False)
+    email: Mapped[str] = mapped_column(nullable=False, unique=True)
+    password: Mapped[str] = mapped_column(nullable=False)
+    phone_number: Mapped[str] = mapped_column(nullable=True)