+++ Changelog.md
@@ -1,5 +1,9 @@
 ### История изменения проекта по semver (https://semver.org/lang/ru/)
 
+# 0.2.0 (29.10.2024)
+- Добавлены model: client
+- Добавлены CRUD-запросы к client в routes
+- client_repo
 
 # 0.1.0 (28.11.2024)
 - В schemas добавлена таблица client